backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
pdfplumber
python-multipart
alembic
python-dotenv
